+++ UVtools.GUI/Properties/Resources.Designer.cs
@@ -673,6 +673,16 @@ namespace UVtools.GUI.Properties {
         /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
+        internal static System.Drawing.Bitmap object_group_16x16 {
+            get {
+                object obj = ResourceManager.GetObject("object-group-16x16", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
         internal static System.Drawing.Bitmap Ok_24x24 {
             get {
                 object obj = ResourceManager.GetObject("Ok-24x24", resourceCulture);